DAOS-10431 build: add ubuntu 22.04 client-only support
This patch migrates Ubuntu support from 20.04 to 22.04 which is the
latest LTS. It also removes support for server-side packages since
Ubuntu support will be limited to the DAOS client going forward.
GHA is also being updated to use 22.04.

To support client-only build, a few extra changes were required:
- Argobots should not be built/required for client-only build
- Header files in the object module needed to be split between
  server (srv_internal.h) and client (obj_internal.h).
- The enumeration code in the object module turns out to use
  server-side routines. The code has thus been split into two
  file: srv_enum.c (server-specific) and obj_enum.c (generic).
  Some server-side code used between server components has been
  exported by adding a new daos_srv/object.h header file.
- The SConscript of the security module has a caching capability
  that breaks the client-only build.
