+++ debian/changelog
@@ -1,3 +1,9 @@
+daos (2.3.100-5) unstable; urgency=medium
+  [ Johann Lombardi]
+  * Only build client-side packages for Ubuntu
+
+ -- Johann Lombardi <johann.lombardi@intel.com> Mon, 25 Apr 2022 09:00:00 +0100
+
 daos (2.3.100-4) unstable; urgency=medium
   [ Lei Huang ]
   * Update libfabric to v1.15.0rc3-1 to include critical performance patches
