+++ debian/control
@@ -16,19 +16,14 @@ Build-Depends: debhelper (>= 10),
                pkg-config,
                python3-dev,
                python3-distro,
-               libabt-dev,
-               libpmemobj-dev (>= 1.11),
                libfuse3-dev,
                libprotobuf-c-dev,
                libjson-c-dev,
-               dpdk-dev,
+               libisal-dev,
                libisal-crypto-dev,
                libcunit1-dev,
                golang-go,
                libboost-dev,
-               libspdk-dev,
-               libipmctl-dev,
-               libraft-dev (= 0.9.1-1401.gc18bcb8),
                python3-tabulate,
                liblz4-dev
 Standards-Version: 4.1.2
@@ -124,15 +119,6 @@ Description: This is the package needed to run the DAOS test suite openmpi tools
  .
  This package contains tests
 
-Package: daos-server-tests
-Architecture: any
-Multi-Arch: same
-Depends: ${shlibs:Depends}, ${misc:Depends},
-         daos-server (= ${binary:Version})
-Description: This is the package needed to run the DAOS test suite (server tests)
- .
- This package contains tests
-
 Package: daos-client
 Section: net
 Architecture: any
@@ -149,22 +135,3 @@ Description: The Distributed Asynchronous Object Storage (DAOS) is an open-sourc
  to optimize performance and cost.
  .
  This package contains the DAOS client.
-
-
-Package: daos-server
-Section: net
-Architecture: any
-Multi-Arch: same
-Depends: ${shlibs:Depends}, ${misc:Depends}, openmpi-bin,
-         ipmctl (>02.00.00.3816), libfabric (>= 1.14.0-1)
-Description: The Distributed Asynchronous Object Storage (DAOS) is an open-source
- software-defined object store designed from the ground up for
- massively distributed Non Volatile Memory (NVM). DAOS takes advantage
- of next generation NVM technology like Storage Class Memory (SCM) and
- NVM express (NVMe) while presenting a key-value storage interface and
- providing features such as transactional non-blocking I/O, advanced
- data protection with self healing on top of commodity hardware, end-
- to-end data integrity, fine grained data control and elastic storage
- to optimize performance and cost.
- .
- This package contains the DAOS server.
