--- debian/daos-server-tests.install
@@ -1,14 +1,0 @@
-usr/bin/jump_pl_map
-usr/bin/ring_pl_map
-usr/bin/evt_ctl
-usr/bin/pl_bench
-usr/bin/rdbt
-usr/bin/smd_ut
-usr/bin/vea_ut
-usr/bin/srv_checksum_tests
-usr/bin/vos_tests
-usr/bin/vea_stress
-usr/bin/vos_perf
-usr/bin/daos_run_io_conf
-usr/bin/obj_ctl
-usr/bin/daos_gen_io_conf
