--- debian/daos-server.install
@@ -1,31 +1,0 @@
-etc/daos/daos_server.yml
-etc/daos/daos_server.yml
-etc/ld.so.conf.d/daos.conf
-etc/daos/vos_size_input.yaml
-# set daos_admin to be setuid root and group daos_server in order to perform privileged tasks
-usr/bin/daos_admin
-# set daos_server to be setgid daos_server in order to invoke daos_admin
-usr/bin/daos_server
-usr/bin/daos_engine
-usr/bin/daos_metrics
-usr/lib64/daos_srv/libcont.so
-usr/lib64/daos_srv/libdtx.so
-usr/lib64/daos_srv/libmgmt.so
-usr/lib64/daos_srv/libobj.so
-usr/lib64/daos_srv/libpool.so
-usr/lib64/daos_srv/librdb.so
-usr/lib64/daos_srv/librdbt.so
-usr/lib64/daos_srv/librebuild.so
-usr/lib64/daos_srv/librsvc.so
-usr/lib64/daos_srv/libsecurity.so
-usr/lib64/daos_srv/libvos_srv.so
-usr/lib64/daos_srv/libvos_size.so
-usr/lib64/daos_srv/libvos.so
-usr/lib64/daos_srv/libbio.so
-usr/lib64/daos_srv/libplacement.so
-usr/lib64/libdaos_common_pmem.so
-usr/share/daos/control/setup_spdk.sh
-usr/lib/systemd/system/daos_server.service
-usr/lib/sysctl.d/10-daos_server.conf
-usr/bin/daos_storage_estimator.py
-usr/lib64/python3.8/site-packages/storage_estimator/*.py
