--- debian/daos-server.postinst
@@ -1,11 +1,0 @@
-#!/bin/sh
-set -e
-
-chown daos_server:daos_server /etc/daos/certs /etc/daos/certs/clients
-chown root:root /etc/daos/daos_server.yml
-# set daos_admin to be setuid root in order to perform privileged tasks
-chown root:daos_server /usr/bin/daos_admin
-# set daos_server to be setgid daos_server in order to invoke daos_admin
-chown root:daos_server /usr/bin/daos_server
-# apply recently installed sysctl conf
-/usr/lib/systemd/systemd-sysctl /usr/lib/sysctl.d/10-daos_server.conf >/dev/null 2>&1 || :
