--- debian/daos-server.preinst
@@ -1,9 +1,0 @@
-#!/bin/sh
-set -e
-
-addgroup --system daos_server
-adduser --system --ingroup daos_server daos_server
-addgroup --system daos_metrics
-usermod -G -a daos_metrics daos_server
-addgroup --system daos_daemons
-usermod -G -a daos_daemons daos_server
