+++ site_scons/prereq_tools/base.py
@@ -737,13 +737,15 @@ class PreReqComponent():
         self.__env.Alias('test', build_dir)
         self._build_targets = []
         check = any(item in BUILD_TARGETS for item in targets)
-        if not check or 'test' in BUILD_TARGETS:
+        if not check:
             self._build_targets.extend(['client', 'server', 'test'])
         else:
             if 'client' in BUILD_TARGETS:
                 self._build_targets.append('client')
             if 'server' in BUILD_TARGETS:
                 self._build_targets.append('server')
+            if 'test' in BUILD_TARGETS:
+                self._build_targets.append('test')
         BUILD_TARGETS.append(build_dir)
 
     def has_source(self, env, *comps, **kw):
@@ -1047,10 +1049,10 @@ class PreReqComponent():
     def load_defaults(self, is_arm):
         """Setup default build parameters"""
         # argobots is not really needed by client but it's difficult to separate
-        common_reqs = ['argobots', 'ofi', 'hwloc', 'mercury', 'boost', 'uuid',
+        common_reqs = ['ofi', 'hwloc', 'mercury', 'boost', 'uuid',
                        'crypto', 'protobufc', 'lz4']
         client_reqs = ['fuse', 'json-c']
-        server_reqs = ['pmdk']
+        server_reqs = ['pmdk', 'argobots']
         test_reqs = ['cmocka']
 
         if not is_arm:
