+++ src/include/daos_srv/daos_engine.h
@@ -732,105 +732,6 @@ int dsc_task_run(tse_task_t *task, tse_task_cb_t retry_cb, void *arg,
 		 int arg_size, bool sync);
 tse_sched_t *dsc_scheduler(void);
 
-typedef int (*iter_copy_data_cb_t)(daos_handle_t ih,
-				   vos_iter_entry_t *it_entry,
-				   d_iov_t *iov_out);
-struct dss_enum_arg {
-	daos_epoch_range_t     *eprs;
-	struct daos_csummer    *csummer;
-	int			eprs_cap;
-	int			eprs_len;
-	int			last_type;	/* hack for tweaking kds_len */
-	iter_copy_data_cb_t	copy_data_cb;
-	/* Buffer fields */
-	union {
-		struct {	/* !fill_recxs */
-			daos_key_desc_t	       *kds;
-			int			kds_cap;
-			int			kds_len;
-			d_sg_list_t	       *sgl;
-			d_iov_t			csum_iov;
-			uint32_t		ec_cell_sz;
-			int			sgl_idx;
-		};
-		struct {	/* fill_recxs && type == S||R */
-			daos_recx_t	       *recxs;
-			int			recxs_cap;
-			int			recxs_len;
-		};
-	};
-	daos_size_t		inline_thres;	/* type == S||R || chk_key2big*/
-	int			rnum;		/* records num (type == S||R) */
-	daos_size_t		rsize;		/* record size (type == S||R) */
-	daos_unit_oid_t		oid;		/* for unpack */
-	uint32_t		fill_recxs:1,	/* type == S||R */
-				chk_key2big:1,
-				need_punch:1,	/* need to pack punch epoch */
-				obj_punched:1,	/* object punch is packed   */
-				size_query:1;	/* Only query size */
-};
-
-struct dtx_handle;
-typedef int (*enum_iterate_cb_t)(vos_iter_param_t *param, vos_iter_type_t type,
-			    bool recursive, struct vos_iter_anchors *anchors,
-			    vos_iter_cb_t pre_cb, vos_iter_cb_t post_cb,
-			    void *arg, struct dtx_handle *dth);
-
-int dss_enum_pack(vos_iter_param_t *param, vos_iter_type_t type, bool recursive,
-		  struct vos_iter_anchors *anchors, struct dss_enum_arg *arg,
-		  enum_iterate_cb_t iter_cb, struct dtx_handle *dth);
-typedef int (*obj_enum_process_cb_t)(daos_key_desc_t *kds, void *ptr,
-				     unsigned int size, void *arg);
-int
-obj_enum_iterate(daos_key_desc_t *kdss, d_sg_list_t *sgl, int nr,
-		 unsigned int type, obj_enum_process_cb_t cb,
-		 void *cb_arg);
-/** Maximal number of iods (i.e., akeys) in dss_enum_unpack_io.ui_iods */
-#define DSS_ENUM_UNPACK_MAX_IODS 16
-
-/**
- * Used by dss_enum_unpack to accumulate recxs that can be stored with a single
- * VOS update.
- *
- * ui_oid and ui_dkey are only filled by dss_enum_unpack for certain
- * enumeration types, as commented after each field. Callers may fill ui_oid,
- * for instance, when the enumeration type is VOS_ITER_DKEY, to pass the object
- * ID to the callback.
- *
- * ui_iods, ui_recxs_caps, and ui_sgls are arrays of the same capacity
- * (ui_iods_cap) and length (ui_iods_len). That is, the iod in ui_iods[i] can
- * hold at most ui_recxs_caps[i] recxs, which have their inline data described
- * by ui_sgls[i]. ui_sgls is optional. If ui_iods[i].iod_recxs[j] has no inline
- * data, then ui_sgls[i].sg_iovs[j] will be empty.
- */
-struct dss_enum_unpack_io {
-	daos_unit_oid_t		 ui_oid;	/**< type <= OBJ */
-	daos_key_t		 ui_dkey;	/**< type <= DKEY */
-	daos_iod_t		*ui_iods;
-	d_iov_t			 ui_csum_iov;
-	/* punched epochs per akey */
-	daos_epoch_t		*ui_akey_punch_ephs;
-	daos_epoch_t		*ui_rec_punch_ephs;
-	daos_epoch_t		**ui_recx_ephs;
-	int			 ui_iods_cap;
-	int			 ui_iods_top;
-	int			*ui_recxs_caps;
-	/* punched epoch for object */
-	daos_epoch_t		ui_obj_punch_eph;
-	/* punched epochs for dkey */
-	daos_epoch_t		ui_dkey_punch_eph;
-	d_sg_list_t		*ui_sgls;	/**< optional */
-	uint32_t		ui_version;
-	uint32_t		ui_type;
-};
-
-typedef int (*dss_enum_unpack_cb_t)(struct dss_enum_unpack_io *io, void *arg);
-
-int
-dss_enum_unpack(daos_unit_oid_t oid, daos_key_desc_t *kds, int kds_num,
-		d_sg_list_t *sgl, d_iov_t *csum, dss_enum_unpack_cb_t cb,
-		void *cb_arg);
-
 d_rank_t dss_self_rank(void);
 
 unsigned int dss_ctx_nr_get(void);
