+++ src/object/SConscript
@@ -8,7 +8,7 @@ def scons():
     env.AppendUnique(LIBPATH=[Dir('.')])
 
     denv = env.Clone()
-    prereqs.require(denv, 'argobots', 'protobufc', 'hwloc')
+    prereqs.require(denv, 'protobufc', 'hwloc')
 
     # Common object code
     common_tgts = denv.SharedObject(['obj_class.c', 'obj_rpc.c', 'obj_task.c',
@@ -27,19 +27,21 @@ def scons():
 
     # generate server module
     senv = denv.Clone()
+    prereqs.require(senv, 'argobots')
     prereqs.require(senv, 'pmdk', headers_only=True)
 
     senv.Append(CPPDEFINES=['-DDAOS_PMEM_BUILD'])
     srv = daos_build.library(senv, 'obj',
                              common_tgts + ['srv_obj.c', 'srv_mod.c',
                                             'srv_obj_remote.c', 'srv_ec.c',
-                                            'srv_obj_migrate.c',
+                                            'srv_obj_migrate.c', 'srv_enum.c',
                                             'srv_cli.c', 'srv_ec_aggregate.c',
                                             'srv_csum.c', 'srv_io_map.c'],
                              install_off="../..")
     senv.Install('$PREFIX/lib64/daos_srv', srv)
 
     if prereqs.test_requested():
+        denv = senv
         SConscript('tests/SConscript', exports='denv')
 
 if __name__ == "SCons.Script":
